libs/data/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neurolab-data"
version = "0.1.0"
description = "Unified data management across local and HPC environments for EEG research"
requires-python = ">=3.10"
license = "MIT"

dependencies = [
    "neurolab-config>=0.1.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
eeg = [
    "mne>=1.6",
    "braindecode>=0.8",
]
processing = [
    "mne>=1.6",
    "scipy>=1.11",
    "joblib>=1.3",
]
torch = [
    "torch>=2.0",
]
dev = [
    "pytest>=7.0",
    "mne>=1.6",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["neurolab.*"]
